Fix BCryptPasswordHasherTests
